BKinD_User/setup/requirements.txt
# python==3.12.2
numpy==1.26.4
pandas==2.2.1
plotly==5.19.0
tqdm==4.66.2
pillow==10.3.0